Property_Managment_System/backend/property_management_system/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1